Resolve a circular import bug introduced by moving get_wavelengths() to instrument.py and get_bin_edges() to binning.py. Easiest way to do this was to move get_ech_slit_indices() to echelle.py, which seems acceptable.
