+++ maven_iuvs/binning.py
@@ -1,5 +1,4 @@
 import numpy as np
-from maven_iuvs.instrument import get_wavelengths
 
 def get_bin_edges(light_fits):
     """
